WIP: testing required

+++ ProjectFiles/Assets/Scripts/Game/RoundManager.cs
@@ -3,6 +3,7 @@ using Network;
 using UnityEngine;
 using Utils;
 
+
 namespace Game
 {
     static class DefaultRoundProperties
@@ -15,7 +16,7 @@ namespace Game
     }
 
     public delegate void TimerOverDelegate();
-
+    
     public class RoundManager
     {
         public bool GameInProgress { get; private set; }
@@ -139,7 +140,7 @@ namespace Game
             RoundEndEvent?.Invoke(roundNumber);
         }
 
-        private void NotifyEliminatePlayers(List<ushort> eliminatedPlayers)
+        private void NotifyEliminatePlayers(List<int> eliminatedPlayers)
         {
             EliminatePlayersEvent?.Invoke(roundNumber, eliminatedPlayers);
         }
